+++ src/pages/FilmView/FilmView.module.css
@@ -13,6 +13,7 @@
 .block {
   display: flex;
   flex-direction: column;
+  gap: 1rem;
   width: calc(50% - 2rem);
 }
